fix  优化us 时间戳的长时间截断误差
Close

+++ components/hal/include/hal_time.h
@@ -37,7 +37,7 @@ void hal_time_delay_us(uint32_t us);
  * @return 返回 DWT->CYCCNT 寄存器实时的存储值
  * @note 使用前必须确保 DWT 初始化成功
  */
-uint32_t hal_get_cycle32(void);
+uint64_t hal_get_cycle32(void);
 
 /**
  * @brief 提供任意合法主频下的us周期
